fix(lib-classifier): delete out-of-bounds drawing marks
- add an out-of-bounds check to the `InteractionLayer` drawing canvas.
- add a unique ID to each mark DOM node.
- when creating a mark, on finish check if the active mark element overlaps the subject image and delete the mark if it does not.
