+++ packages/lib-classifier/src/plugins/drawingTools/components/Mark/Mark.js
@@ -56,6 +56,7 @@ const Mark = forwardRef(function Mark(
     children,
     disabled = false,
     dragging = false,
+    id,
     isActive = false,
     label,
     mark,
@@ -153,6 +154,7 @@ const Mark = forwardRef(function Mark(
   return (
     <StyledGroup
       {...mainStyle}
+      id={id}
       data-testid="mark-mark"
       aria-disabled={disabled ? 'true' : 'false'}
       aria-label={label}
@@ -177,6 +179,7 @@ const Mark = forwardRef(function Mark(
 Mark.propTypes = {
   dragging: PropTypes.bool,
   children: PropTypes.node.isRequired,
+  id: PropTypes.string,
   isActive: PropTypes.bool,
   label: PropTypes.string.isRequired,
   onDelete: PropTypes.func,
